BackEnd/Python/python-django-pip-djangoorm-postgresql/requirements.txt
Django==5.2.12
psycopg[binary]==3.3.3
python-dotenv==1.2.2
PyJWT==2.12.1
Pillow==10.4.0
